+++ packages/core/src/index.ts
@@ -1,7 +1,18 @@
 // SECTION: Imports
 import logger from './logger';
+
+import {
+    APIResponse,
+    APIResponseSchema,
+    LoginUserBody,
+    LoginUserBodySchema,
+    RegisterUserBody,
+    RegisterUserBodySchema
+} from './schema/auth';
 // !SECTION: Imports
 
 // SECTION: Exports
 export { logger };
+export { APIResponseSchema, LoginUserBodySchema, RegisterUserBodySchema };
+export type { APIResponse, LoginUserBody, RegisterUserBody };
 // !SECTION: Exports